Add s3 configuration to use in prod
